+++ seeds-game-model/src/main/java/com/seeds/game/dto/response/NftPublicBackpackResp.java
@@ -42,7 +42,7 @@ public class NftPublicBackpackResp implements Serializable {
     private Long itemId;
 
     @ApiModelProperty("游戏那边传的。大区+game+流水")
-    private Integer autoId;
+    private Long autoId;
 
     @ApiModelProperty("用户id")
     private Long userId;
